backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
mysqlclient
python-dotenv
celery
django-celery-beat
redis
python-dateutil
requests